tests - move tests to test folder
