+++ tests/AssetsContractController.test.ts
@@ -1,4 +1,4 @@
-import { AssetsContractController } from './AssetsContractController';
+import { AssetsContractController } from '../src/AssetsContractController';
 const HttpProvider = require('ethjs-provider-http');
 const MAINNET_PROVIDER = new HttpProvider('https://mainnet.infura.io');
 const GODSADDRESS = '0x6EbeAf8e8E946F0716E6533A6f2cefc83f60e8Ab';
