+++ tests/util.test.ts
@@ -1,4 +1,4 @@
-import * as util from './util';
+import * as util from '../src/util';
 import { mock } from 'fetch-mock';
 
 const { BN } = require('ethereumjs-util');
